Day 11 part 2 completed. Many changes from part 1 because not sustainable to keep a literal representation of the universe with the expansion increments. New method keeps track of offsets horizontally and vertically for the different indices instead.

+++ src/Day11/Galaxy.java
@@ -1,19 +1,19 @@
 package Day11;
 
 public class Galaxy {
-    private int row;
-    private int col;
+    private long row;
+    private long col;
 
-    public Galaxy(int row, int col) {
+    public Galaxy(long row, long col) {
         this.row = row;
         this.col = col;
     }
 
-    public int getRow() {
+    public long getRow() {
         return row;
     }
 
-    public int getCol() {
+    public long getCol() {
         return col;
     }
 }
